requirements.txt
flask
flask-restx
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
python-dotenv
flask-marshmallow
marshmallow-SQLAlchemy
email-validator
Flask-WTF
gunicorn